Adds support for child IoC container+
MvxSimpleIoCContainer is only a wrapper around MvxIoCContainer now as the container can't be a singleton anymore.

Breaking changes: the previously virtual methods InjectProperties and GetIoCParameterValues from MvxSimpleIoCContainer have moved to MvxIoCContainer
