+++ MvvmCross/Platform/Platform/IoC/MvxIoCProviderExtensions.cs
@@ -1,0 +1,10 @@
+﻿namespace MvvmCross.Platform.IoC
+{
+    public static class MvxIoCProviderExtensions
+    {
+        public static IMvxIoCProvider CreateChildContainer(this IMvxIoCProvider iocProvider)
+        {
+            return new MvxIoCContainer(iocProvider);
+        }
+    }
+}
