+++ MvvmCross/Platform/Platform/MvvmCross.Platform.csproj
@@ -132,6 +132,8 @@
     <Compile Include="Logging\MvxLogExtensions.cs" />
     <Compile Include="Logging\IMvxLogProvider.cs" />
     <Compile Include="Logging\MvxLogProviderType.cs" />
+    <Compile Include="IoC\MvxIoCContainer.cs" />
+    <Compile Include="IoC\MvxIoCProviderExtensions.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
